+++ lib/src/lib.rs
@@ -49,6 +49,7 @@ pub mod merge;
 pub mod merged_tree;
 pub mod op_heads_store;
 pub mod op_store;
+pub mod op_walk;
 pub mod operation;
 #[allow(missing_docs)]
 pub mod protos;
